DESKTOP-1051: Fix pending client connections breaking the dummy server boot.
NOTE: This isn't a perfect implementation. In an ideal world, we would take every pending client request and pass them onto real server threads, and then connect the dummy client after all the pending client requests are dealt with. Right now, we're closing all pending client connections and then starting the server. This means all clients will get an EOF or similar error if they try to connect before the server is ready, and will need to retry.

This might not be a big deal because often clients have a timeout / retry loop waiting for the server to come up anyway and if you try to connect before the server is ready, its reasonable to expect errors.

To reach the ideal world we'd need some way of the server identifying a dummy client connection and keeping it open. This would require a change to the protocol which may interfere with the generic named pipe solution, which is why i'm committing the basic, reliable but somewhat awkward implementation here.

Signed-off-by: Mike Parker <michael.parker@docker.com>